requirements.txt
cat > requirements.txt << 'EOF'
python-dotenv>=1.0.0
pandas>=2.0.0
tabulate>=0.9.0
colorama>=0.4.6
pydantic>=2.5.0
aiohttp>=3.9.0
requests>=2.31.0
EOF

# Install the packages
pip install -r requirements.txt